+++ priv/repo/migrations/20191119224938_todo_is_done_field_can_no_longer_be_null.exs
@@ -1,0 +1,11 @@
+defmodule Macroy.Repo.Migrations.TodoIsDoneFieldCanNoLongerBeNull do
+  use Ecto.Migration
+
+  def up do
+    alter table("todos"), do: modify :is_done, :boolean, null: false
+  end
+
+  def down do
+    alter table("todos"), do: modify :is_done, :boolean, null: true
+  end
+end