+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.2.1] - 2019-11-19
 ### Added
 - Test coverage on OrgFile.read/1 will check to see if `todo.is_done` is filled
   in based on whether `todo.closed_on` is found. 
@@ -102,10 +103,11 @@ work, and can be closed.
 - Link from OrgFile index to home.
 - Link from home to OrgFile index.
 
-[Unreleased]: https://github.com/brotherjack/macroy/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/brotherjack/macroy/compare/v1.2.1...HEAD
 [0.0.1]: https://github.com/brotherjack/macroy/releases/tag/v0.0.1
 [0.0.2]: https://github.com/brotherjack/macroy/compare/v0.0.1...v0.0.2
 [0.1.0]: https://github.com/brotherjack/macroy/compare/v0.0.2...v0.1.0
 [1.0.0]: https://github.com/brotherjack/macroy/compare/v0.1.0...v1.0.0
 [1.1.0]: https://github.com/brotherjack/macroy/compare/v1.0.0...v1.1.0
 [1.2.0]: https://github.com/brotherjack/macroy/compare/v1.1.0...v1.2.0
+[1.2.1]: https://github.com/brotherjack/macroy/compare/v1.2.0...v1.2.1